[kernel emul] move definition of bsg_print to cpp file
